test: extract and prune core unit tests
Move core unit tests under tests/crate and remove ten redundant or
ineffective tests: seven with stronger coverage elsewhere and three
that reconstructed summary-table formatting without exercising the
production path.
