+++ src/main.rs
@@ -228,466 +228,5 @@ fn main() {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::cli::Flags;
-    use crate::text_processing::{
-        BinaryReason, ProcessedFile, read_classify_and_decode,
-    };
-    use clap::Parser;
-    use std::fs;
-    use std::str::FromStr;
-
-    fn create_test_config(toml_content: &str) -> Params {
-        let config = Config::builder()
-            .add_source(config::File::from_str(
-                toml_content,
-                config::FileFormat::Toml,
-            ))
-            .build()
-            .unwrap();
-        config.into()
-    }
-
-    #[test]
-    fn test_exclude_takes_precedence_over_extend_exclude() {
-        // Setup CLI args with both exclude and extend-exclude
-        let args = Flags::parse_from([
-            "program",
-            "--exclude",
-            "*.txt",
-            "--extend-exclude",
-            "*.md",
-        ]);
-
-        // Create config with both exclude and extend-exclude
-        let config = create_test_config(
-            r#"
-            extend_exclude = ["*.rs"]
-            exclude = ["*.toml"]
-        "#,
-        );
-
-        let params = Params::from_args_and_config(&args, config);
-
-        // Verify that extend_exclude is None when exclude is set
-        assert!(params.extend_exclude.is_none());
-        // Verify that exclude contains only CLI patterns
-        assert_eq!(params.exclude, Some(vec!["*.txt".to_string()]));
-    }
-
-    #[test]
-    fn test_cli_exclude_overrides_config_exclude() {
-        let args = Flags::parse_from([
-            "program",
-            "--exclude",
-            "*.txt",
-            "--exclude",
-            "*.md",
-        ]);
-
-        let config = create_test_config(
-            r#"
-            exclude = ["*.toml", "*.rs"]
-        "#,
-        );
-
-        let params = Params::from_args_and_config(&args, config);
-
-        assert_eq!(
-            params.exclude,
-            Some(vec!["*.txt".to_string(), "*.md".to_string()])
-        );
-    }
-
-    #[test]
-    fn test_extend_exclude_combines_cli_and_config() {
-        let args = Flags::parse_from([
-            "program",
-            "--extend-exclude",
-            "*.txt",
-            "--extend-exclude",
-            "*.md",
-        ]);
-
-        let config = create_test_config(
-            r#"
-            extend_exclude = ["*.toml", "*.rs"]
-        "#,
-        );
-
-        let params = Params::from_args_and_config(&args, config);
-
-        assert_eq!(
-            params.extend_exclude,
-            Some(vec![
-                "*.txt".to_string(),
-                "*.md".to_string(),
-                "*.toml".to_string(),
-                "*.rs".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn test_config_exclude_disables_extend_exclude() {
-        let args = Flags::parse_from(["program", "--extend-exclude", "*.txt"]);
-
-        let config = create_test_config(
-            r#"
-            exclude = ["*.toml"]
-            extend_exclude = ["*.rs"]
-        "#,
-        );
-
-        let params = Params::from_args_and_config(&args, config);
-
-        assert!(params.extend_exclude.is_none());
-        assert_eq!(params.exclude, Some(vec!["*.toml".to_string()]));
-    }
-
-    #[test]
-    fn test_no_exclude_patterns() {
-        let args = Flags::parse_from(["program"]);
-        let config = create_test_config("");
-
-        let params = Params::from_args_and_config(&args, config);
-
-        assert!(params.exclude.is_none());
-        assert!(params.extend_exclude.is_none());
-    }
-
-    #[test]
-    fn test_model_parsing() {
-        use std::str::FromStr;
-
-        // Test valid model parsing
-        let model = Model::from_str("gpt5").unwrap();
-        assert!(matches!(model, Model::GPT5));
-
-        // Test invalid model
-        let invalid = Model::from_str("invalid_model");
-        assert!(invalid.is_err());
-    }
-
-    #[test]
-    fn test_summary_table_formatting() {
-        let summary_data = vec![
-            SummaryTable {
-                metric: "Files:".to_string(),
-                value: "10".to_string(),
-            },
-            SummaryTable {
-                metric: "Size:".to_string(),
-                value: "1024".to_string(),
-            },
-        ];
-
-        let table = Table::new(summary_data)
-            .with(Remove::row(Rows::first()))
-            .with(Style::empty())
-            .with(Modify::list(Columns::first(), Alignment::right()))
-            .to_string();
-
-        // Check that the table is formatted correctly
-        assert!(table.contains("Files:"));
-        assert!(table.contains("10"));
-        assert!(table.contains("Size:"));
-        assert!(table.contains("1024"));
-    }
-
-    #[test]
-    fn test_version_flag() {
-        let args = Flags::parse_from(["bundlerepo", "--version"]);
-        assert!(args.version);
-    }
-
-    #[test]
-    fn test_model_parsing_error() {
-        let result = Model::from_str("invalid_model");
-        assert!(result.is_err());
-        let err = result.unwrap_err();
-        println!("Error: {}", err);
-        assert!(err.contains("Unsupported model"));
-    }
-
-    #[test]
-    fn test_repo_clone_error() {
-        let temp_dir = tempdir().unwrap();
-        let args = Flags::parse_from(["bundlerepo", "invalid_repo"]);
-        let config = Params::default();
-        let params = Params::from_args_and_config(&args, config);
-        let result =
-            repo::clone_repo(&params, "invalid_repo", None, temp_dir.path());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_current_directory_check() {
-        let temp_dir = tempdir().unwrap();
-        let params = Params::default();
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-        let result = repo::check_current_directory(&params);
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("Not a git repository")
-        );
-    }
-
-    #[test]
-    fn test_xml_output_error() {
-        let temp_dir = tempdir().unwrap();
-        let params = Params {
-            output_file: Some("/nonexistent/directory/output.xml".to_string()),
-            ..Params::default()
-        };
-        let file_tree = filelist::group_files_by_directory(vec![]);
-        let model = Model::GPT4o;
-        let tokenizer = model.to_tokenizer().unwrap();
-        let result = xml_output::output_repo_as_xml(
-            &params,
-            file_tree,
-            temp_dir.path(),
-            &tokenizer,
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_summary_table_with_clipboard() {
-        let summary_data = vec![
-            SummaryTable {
-                metric: "Files:".to_string(),
-                value: "10".to_string(),
-            },
-            SummaryTable {
-                metric: "Size:".to_string(),
-                value: "1024".to_string(),
-            },
-        ];
-
-        let table = Table::new(summary_data)
-            .with(Remove::row(Rows::first()))
-            .with(Style::empty())
-            .with(Modify::list(Columns::first(), Alignment::right()))
-            .to_string();
-
-        // Check that the table is formatted correctly with clipboard settings
-        assert!(table.contains("Files:"));
-        assert!(table.contains("10"));
-        assert!(table.contains("Size:"));
-        assert!(table.contains("1024"));
-    }
-
-    #[test]
-    fn test_tokenizer_creation() {
-        // Test successful tokenizer creation
-        let model = Model::GPT4o;
-        let tokenizer_result = model.to_tokenizer();
-        assert!(tokenizer_result.is_ok());
-
-        // Test that we can use the tokenizer
-        let tokenizer = tokenizer_result.unwrap();
-        let count = tokenizer.count_tokens("test string").unwrap();
-        assert!(count > 0);
-    }
-
-    #[test]
-    fn test_empty_input_tokenization() {
-        let model = Model::GPT4o;
-        let tokenizer = model.to_tokenizer().unwrap();
-        let count = tokenizer.count_tokens("").unwrap();
-        // Just verify the tokenization succeeded
-        assert_eq!(count, 0);
-    }
-
-    #[test]
-    fn test_model_display_in_summary() {
-        let model = Model::GPT4o;
-        let summary_data = vec![SummaryTable {
-            metric: format!("Token count ({}):", model.display_name()),
-            value: "100".to_string(),
-        }];
-
-        let table = Table::new(summary_data)
-            .with(Remove::row(Rows::first()))
-            .with(Style::empty())
-            .with(Modify::list(Columns::first(), Alignment::right()))
-            .to_string();
-
-        assert!(
-            table
-                .contains(&format!("Token count ({}):", model.display_name()))
-        );
-        assert!(table.contains("100"));
-    }
-
-    #[test]
-    fn test_invalid_model_parsing() {
-        // Test direct model parsing without CLI
-        let result = "invalid_model".parse::<Model>();
-        assert!(result.is_err());
-        let err = result.unwrap_err();
-        assert!(err.contains("Unsupported model"));
-    }
-
-    #[test]
-    fn test_invalid_tokenizer_creation() {
-        // This is a bit of a contrived test since our current models all create valid tokenizers
-        // but it ensures we handle empty input correctly
-        let model = Model::GPT4o;
-        let tokenizer = model.to_tokenizer().unwrap();
-        let count = tokenizer.count_tokens("test string").unwrap();
-        assert!(count > 0);
-    }
-
-    #[test]
-    fn test_utf8_flag_overrides_config_false() {
-        let config = create_test_config(
-            r#"
-            utf8 = false
-        "#,
-        );
-        let args = Flags::parse_from(["program", "--utf8"]);
-        let params = Params::from_args_and_config(&args, config);
-        assert!(params.utf8);
-    }
-
-    #[test]
-    fn test_no_utf8_flag_overrides_config_true() {
-        let config = create_test_config(
-            r#"
-            utf8 = true
-        "#,
-        );
-        let args = Flags::parse_from(["program", "--no-utf8"]);
-        let params = Params::from_args_and_config(&args, config);
-        assert!(!params.utf8);
-    }
-
-    #[test]
-    fn test_config_utf8_used_when_no_flags() {
-        let config = create_test_config(
-            r#"
-            utf8 = true
-        "#,
-        );
-        let args = Flags::parse_from(["program"]);
-        let params = Params::from_args_and_config(&args, config);
-        assert!(params.utf8);
-    }
-
-    #[test]
-    fn test_default_utf8_when_no_config_or_flags() {
-        let config = create_test_config("");
-        let args = Flags::parse_from(["program"]);
-        let params = Params::from_args_and_config(&args, config);
-        assert!(!params.utf8);
-    }
-
-    #[test]
-    fn test_utf8_precedence_controls_utf16_conversion() {
-        let temp_dir = tempdir().unwrap();
-        let fixture_path = temp_dir.path().join("utf-16le.txt");
-        fs::write(
-            &fixture_path,
-            include_bytes!("../tests/fixtures/encodings/utf-16le.txt"),
-        )
-        .unwrap();
-        let cases: [(&str, &[&str], bool); 6] = [
-            ("utf8 = false", &["program"], false),
-            ("utf8 = false", &["program", "--utf8"], true),
-            ("utf8 = false", &["program", "--no-utf8"], false),
-            ("utf8 = true", &["program"], true),
-            ("utf8 = true", &["program", "--utf8"], true),
-            ("utf8 = true", &["program", "--no-utf8"], false),
-        ];
-
-        for (config, arguments, expected_utf8) in cases {
-            let args = Flags::parse_from(arguments);
-            let params = Params::from_args_and_config(
-                &args,
-                create_test_config(config),
-            );
-            assert_eq!(params.utf8, expected_utf8);
-
-            let mut timings = timings::ProcessingTimings::default();
-            let processed = read_classify_and_decode(
-                &fixture_path,
-                params.utf8,
-                &mut timings,
-            )
-            .unwrap();
-            if expected_utf8 {
-                match processed {
-                    ProcessedFile::Text(decoded) => {
-                        assert!(decoded.text.starts_with("UTF-16 text"));
-                        assert_eq!(timings.transcoded_files, 1);
-                    }
-                    ProcessedFile::Binary(reason) => {
-                        panic!("expected decoded UTF-16, got {reason:?}")
-                    }
-                }
-            } else {
-                assert_eq!(
-                    processed,
-                    ProcessedFile::Binary(
-                        BinaryReason::Utf16ConversionDisabled("UTF-16LE")
-                    )
-                );
-                assert_eq!(timings.transcoded_files, 0);
-            }
-        }
-    }
-
-    #[test]
-    fn test_gzip_cli_and_config_precedence() {
-        let cases = [
-            ("", vec!["program"], false, 6),
-            ("gzip = false\ngzip_level = 9", vec!["program"], false, 9),
-            ("", vec!["program", "-z"], true, 6),
-            ("gzip = true", vec!["program"], true, 6),
-            ("gzip = true\ngzip_level = 8", vec!["program"], true, 8),
-            (
-                "gzip = false\ngzip_level = 9",
-                vec!["program", "-z"],
-                true,
-                9,
-            ),
-            (
-                "gzip = true\ngzip_level = 8",
-                vec!["program", "-z=3"],
-                true,
-                3,
-            ),
-            (
-                "gzip = true\ngzip_level = 8",
-                vec!["program", "--no-gzip"],
-                false,
-                8,
-            ),
-        ];
-
-        for (config, arguments, expected_gzip, expected_level) in cases {
-            let args = Flags::parse_from(arguments);
-            let params = Params::from_args_and_config(
-                &args,
-                create_test_config(config),
-            );
-            assert_eq!(params.gzip, expected_gzip);
-            assert_eq!(params.gzip_level, expected_level);
-        }
-    }
-
-    #[test]
-    fn test_no_gzip_allows_clipboard_override() {
-        let config = create_test_config("gzip = true\ngzip_level = 9");
-        let args = Flags::parse_from(["program", "--no-gzip", "--clipboard"]);
-        let params = Params::from_args_and_config(&args, config);
-        assert!(!params.gzip);
-        assert!(params.clipboard);
-    }
-}
+#[path = "../tests/crate/app.rs"]
+mod tests;
